Move start_preparing_data_collection_then_do_plan into common
